Refactor: Replace WithContext with Step and update README badges
