+++ .github/workflows/main.yaml
@@ -38,7 +38,7 @@ jobs:
           JRELEASER_GPG_PASSPHRASE: ${{ secrets.JRELEASER_GPG_PASSPHRASE }}
           JRELEASER_GPG_SECRET_KEY: ${{ secrets.JRELEASER_GPG_SECRET_KEY }}
           JRELEASER_GPG_PUBLIC_KEY: ${{ secrets.JRELEASER_GPG_PUBLIC_KEY }}
-        run: ./gradlew publish jreleaserDeploy
+        run: ./gradlew publish jreleaserDeploy --stacktrace
       - name: Test Report
         uses: dorny/test-reporter@v1
         if: success() || failure()
