+++ src/main/kotlin/io/github/ludorival/kotlintdd/AssumeActAssert.kt
@@ -1,6 +1,6 @@
 package io.github.ludorival.kotlintdd
 
-abstract class AssumeActAssert<R1 : WithContext, R2 : WithContext, R3 : WithContext>(
+abstract class AssumeActAssert<R1 : Step, R2 : Step, R3 : Step>(
     assumption: R1,
     action: R2,
     assertion: R3
@@ -10,14 +10,14 @@ abstract class AssumeActAssert<R1 : WithContext, R2 : WithContext, R3 : WithCont
 }
 
 object SimpleAssumeActAssert :
-    AssumeActAssert<WithContext, WithContext, WithContext>(WithContext(), WithContext(), WithContext())
+    AssumeActAssert<Step, Step, Step>(Step(), Step(), Step())
 
-infix fun <T, V, R1 : WithContext, R2 : WithContext, R3 : WithContext> BasePattern.AssumptionContext<T, R1, R2, R3>.act(
+infix fun <T, V, R1 : Step, R2 : Step, R3 : Step> BasePattern.AssumptionContext<T, R1, R2, R3>.act(
     block: R2.(T) -> V
 ) =
     chainAct("ACT", block)
 
-infix fun <T, V, R1 : WithContext, R2 : WithContext, R3 : WithContext> BasePattern.ActContext<T, R1, R2, R3>.assert(
+infix fun <T, V, R1 : Step, R2 : Step, R3 : Step> BasePattern.ActContext<T, R1, R2, R3>.assert(
     block: R3.(T) -> V
 ) =
     chainAssert("ASSERT", block)
