+++ src/main/kotlin/io/github/ludorival/kotlintdd/GivenWhenThen.kt
@@ -1,7 +1,7 @@
 package io.github.ludorival.kotlintdd
 
 @Suppress("FunctionNaming")
-abstract class GivenWhenThen<R1 : WithContext, R2 : WithContext, R3 : WithContext>(
+abstract class GivenWhenThen<R1 : Step, R2 : Step, R3 : Step>(
     assumption: R1,
     action: R2,
     assertion: R3
@@ -11,21 +11,21 @@ abstract class GivenWhenThen<R1 : WithContext, R2 : WithContext, R3 : WithContex
 }
 
 object SimpleGivenWhenThen :
-    GivenWhenThen<WithContext, WithContext, WithContext>(WithContext(), WithContext(), WithContext())
+    GivenWhenThen<Step, Step, Step>(Step(), Step(), Step())
 
 @Suppress("FunctionNaming")
-infix fun <T, V, R1 : WithContext, R2 :
-WithContext, R3 : WithContext> BasePattern.AssumptionContext<T, R1, R2, R3>.`when`(
+infix fun <T, V, R1 : Step, R2 :
+Step, R3 : Step> BasePattern.AssumptionContext<T, R1, R2, R3>.`when`(
     block: R2.(T) -> V
 ) =
     chainAct("WHEN", block)
 
-infix fun <T, V, R1 : WithContext, R2 : WithContext,
-    R3 : WithContext> BasePattern.ActContext<T, R1, R2, R3>.then(block: R3.(T) -> V) =
+infix fun <T, V, R1 : Step, R2 : Step,
+    R3 : Step> BasePattern.ActContext<T, R1, R2, R3>.then(block: R3.(T) -> V) =
     chainAssert("THEN", block)
 
-infix fun <T, V, R1 : WithContext, R2 : WithContext,
-    R3 : WithContext> BasePattern.AssumptionContext<T, R1, R2, R3>.then(
+infix fun <T, V, R1 : Step, R2 : Step,
+    R3 : Step> BasePattern.AssumptionContext<T, R1, R2, R3>.then(
     block: R3.(T) -> V
 ) =
     chainAssert("THEN", block)
