+++ src/main/kotlin/io/github/ludorival/kotlintdd/coroutine/CoAssumeActAssert.kt
@@ -1,8 +1,8 @@
 package io.github.ludorival.kotlintdd.coroutine
 
-import io.github.ludorival.kotlintdd.WithContext
+import io.github.ludorival.kotlintdd.Step
 
-abstract class CoAssumeActAssert<R1 : WithContext, R2 : WithContext, R3 : WithContext>(
+abstract class CoAssumeActAssert<R1 : Step, R2 : Step, R3 : Step>(
     assumption: R1, action: R2, assertion: R3
 ) : CoBasePattern<R1, R2, R3>(assumption, action, assertion) {
     suspend fun <R> assume(block: suspend R1.() -> R) = assume("ASSUME", block)
@@ -10,18 +10,18 @@ abstract class CoAssumeActAssert<R1 : WithContext, R2 : WithContext, R3 : WithCo
 }
 
 object CoSimpleAssumeActAssert :
-    CoAssumeActAssert<WithContext, WithContext, WithContext>(WithContext(), WithContext(), WithContext())
+    CoAssumeActAssert<Step, Step, Step>(Step(), Step(), Step())
 
 suspend infix fun <T, V,
-    R1 : WithContext,
-    R2 : WithContext,
-    R3 : WithContext> CoBasePattern.AssumptionContext<T, R1, R2, R3>.act(
+    R1 : Step,
+    R2 : Step,
+    R3 : Step> CoBasePattern.AssumptionContext<T, R1, R2, R3>.act(
     block: suspend R2.(T) -> V
 ) = chainAct("ACT", block)
 
 suspend infix fun <T, V,
-    R1 : WithContext,
-    R2 : WithContext,
-    R3 : WithContext> CoBasePattern.ActContext<T, R1, R2, R3>.assert(
+    R1 : Step,
+    R2 : Step,
+    R3 : Step> CoBasePattern.ActContext<T, R1, R2, R3>.assert(
     block: suspend R3.(T) -> V
 ) = chainAssert("ASSERT", block)
