+++ src/main/kotlin/io/github/ludorival/kotlintdd/coroutine/CoBasePattern.kt
@@ -1,16 +1,16 @@
 package io.github.ludorival.kotlintdd.coroutine
 
 import io.github.ludorival.kotlintdd.Context
-import io.github.ludorival.kotlintdd.WithContext
+import io.github.ludorival.kotlintdd.Step
 
 
-open class CoBasePattern<R1 : WithContext, R2 : WithContext, R3 : WithContext>(
+open class CoBasePattern<R1 : Step, R2 : Step, R3 : Step>(
     private val assumption: R1,
     private val action: R2,
     private val assertion: R3
 ) {
 
-    class AssumptionContext<T, R1 : WithContext, R2 : WithContext, R3 : WithContext> internal constructor(
+    class AssumptionContext<T, R1 : Step, R2 : Step, R3 : Step> internal constructor(
         private val pattern: CoBasePattern<R1, R2, R3>,
         key: String,
         result: T
@@ -29,7 +29,7 @@ open class CoBasePattern<R1 : WithContext, R2 : WithContext, R3 : WithContext>(
 
     }
 
-    class ActContext<T, R1 : WithContext, R2 : WithContext, R3 : WithContext> internal constructor(
+    class ActContext<T, R1 : Step, R2 : Step, R3 : Step> internal constructor(
         private val pattern: CoBasePattern<R1, R2, R3>,
         key: String,
         result: T
@@ -44,7 +44,7 @@ open class CoBasePattern<R1 : WithContext, R2 : WithContext, R3 : WithContext>(
 
     }
 
-    class AssertContext<T, R1 : WithContext, R2 : WithContext, R3 : WithContext> internal constructor(
+    class AssertContext<T, R1 : Step, R2 : Step, R3 : Step> internal constructor(
         private val pattern: CoBasePattern<R1, R2, R3>,
         key: String,
         result: T
