+++ src/test/kotlin/io/github/ludorival/kotlintdd/dsl/Assertion.kt
@@ -1,10 +1,10 @@
 package io.github.ludorival.kotlintdd.dsl
 
-import io.github.ludorival.kotlintdd.WithContext
+import io.github.ludorival.kotlintdd.Step
 import org.junit.jupiter.api.Assertions
 
 @Suppress("VariableNaming")
-class Assertion : WithContext() {
+class Assertion : Step() {
 
     infix fun <T> T.`should be equal to`(expected: T) = Assertions.assertEquals(expected, this)
